requirements.txt
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0
gradio>=4.0.0
pillow>=10.0.0
numpy>=1.24.0
tqdm>=4.65.0
torchmetrics>=1.0.0
pandas>=2.0.0
openpyxl>=3.1.0 